+++ go.mod
@@ -107,7 +107,6 @@ require (
 )
 
 require (
-	github.com/gofrs/uuid v3.3.0+incompatible // indirect
 	go.uber.org/atomic v1.7.0 // indirect
 	go.uber.org/multierr v1.6.0 // indirect
 )
