+++ npm/pkg/dataplane/ipsets/ipsetmanager_linux.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/parse"
 	"github.com/Azure/azure-container-networking/npm/util"
 	npmerrors "github.com/Azure/azure-container-networking/npm/util/errors"
-	ioutil "github.com/Azure/azure-container-networking/npm/util/osutil_linux"
+	"github.com/Azure/azure-container-networking/npm/util/ioutil"
 	"k8s.io/klog"
 )
