+++ npm/util/ioutil/current-azure-chains.go
