+++ npm/util/ioutil/current-azure-chains_test.go
