+++ npm/util/ioutil/restore.go
