+++ npm/util/ioutil/restore_test.go
